requirements.txt
Flask==2.3.3
Jinja2==3.1.2
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pdfplumber==0.10.3
reportlab==4.0.7
Markdown==3.5.1